+++ typings/luxon.d.ts
@@ -5,6 +5,7 @@ declare module "luxon" {
     static fromHTTP(text: string, options?: Object): DateTime;
     static fromISO(text: string, options?: Object): DateTime;
     static fromRFC2822(text: string, options?: Object): DateTime;
+    static fromString(text: string, format: string, options?: Object): DateTime;
 
     readonly invalidReason: string | null;
     readonly isValid: boolean;